+++ packages/xrpl/test/models/utils.test.ts
@@ -17,6 +17,7 @@ import { isFlagEnabled } from '../../src/models/utils'
 import {
   setTransactionFlagsToNumber,
   parseAccountRootFlags,
+  parseTransactionFlags,
 } from '../../src/models/utils/flags'
 
 /**
@@ -206,6 +207,61 @@ describe('Models Utils', function () {
       assert.isUndefined(parsed.lsfDisallowIncomingPayChan)
       assert.isUndefined(parsed.lsfDisallowIncomingTrustline)
       assert.isUndefined(parsed.lsfAllowTrustLineClawback)
+    })
+
+    it('parseTransactionFlags all enabled', function () {
+      const tx: PaymentChannelClaim = {
+        Account: 'r...',
+        TransactionType: 'PaymentChannelClaim',
+        Channel:
+          'C1AE6DDDEEC05CF2978C0BAD6FE302948E9533691DC749DCDD3B9E5992CA6198',
+        Flags: {
+          tfRenew: true,
+          tfClose: false,
+        },
+      }
+
+      const expected = {
+        tfRenew: true,
+      }
+
+      const flagsMap = parseTransactionFlags(tx)
+      assert.notStrictEqual(flagsMap, expected)
+    })
+
+    it('parseTransactionFlags all false', function () {
+      const tx: PaymentChannelClaim = {
+        Account: 'r...',
+        TransactionType: 'PaymentChannelClaim',
+        Channel:
+          'C1AE6DDDEEC05CF2978C0BAD6FE302948E9533691DC749DCDD3B9E5992CA6198',
+        Flags: {
+          tfRenew: false,
+          tfClose: false,
+        },
+      }
+
+      const expected = {}
+
+      const flagsMap = parseTransactionFlags(tx)
+      assert.notStrictEqual(flagsMap, expected)
+    })
+
+    it('parseTransactionFlags flag is already numeric', function () {
+      const tx: PaymentChannelClaim = {
+        Account: 'r...',
+        TransactionType: 'PaymentChannelClaim',
+        Channel:
+          'C1AE6DDDEEC05CF2978C0BAD6FE302948E9533691DC749DCDD3B9E5992CA6198',
+        Flags: 65536,
+      }
+
+      const expected = {
+        tfRenew: true,
+      }
+
+      const flagsMap = parseTransactionFlags(tx)
+      assert.notStrictEqual(flagsMap, expected)
     })
   })
 })